renamed factory and remove fixture tag
